requirements.txt
fastapi
uvicorn
torch
torchvision
ftfy
regex
tqdm
requests
Pillow
